# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API документация к решению команды MISIS Banach Space для детекции точек незаконной торгволи",
        "title": "ЛЦТ-2023 видео-детекция незаконной торговли API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/videos": {
            "get": {
                "description": "Получение всех видео по фильтру с возможностью пагинации (полтзователи могут получать только видео из своих групп)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Получение всех видео по фильтру",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "status",
                            "groupId"
                        ],
                        "type": "string",
                        "default": "groupId",
                        "description": "Фильтр поиска",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "0",
                        "description": "Значение фильтра",
                        "name": "value",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список видео",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Video"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание видео с указанными данными (доступно только для администраторов)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Создание видео",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Видео файл",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название видео",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID группы, к которой принадлежит видео, 0 - для всех",
                        "name": "groupId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Видео успешно создано",
                        "schema": {
                            "$ref": "#/definitions/model.VideoCreate"
                        }
                    },
                    "400": {
                        "description": "Ошибка при создании видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/many": {
            "post": {
                "description": "Создание нескольких видео из архива с указанными данными (доступно только для администраторов)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Создание нескольких видео из архива",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Архив с видео",
                        "name": "archive",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название видео (будет добавлено к названию каждого видео)",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "ID группы, к которой принадлежит видео (все видео будут добавлены в эту группу), 0 - для всех",
                        "name": "groupId",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Видео успешно созданы",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ошибка при создании видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/updateGroup": {
            "post": {
                "description": "Добавление/удаление видео из гурппы по id (доступно только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Добавление/удаление видео из группы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VideoGroupUpdate"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Видео успешно обновлено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка при обновлении видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}": {
            "get": {
                "description": "Получение видео по id (пользователи могут получать только видео из своих групп)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Получение видео по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id видео",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Видео",
                        "schema": {
                            "$ref": "#/definitions/model.Video"
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление видео по id (доступно только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Удаление видео",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id видео",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Видео успешно удалено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка при удалении видео",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/videos/{id}/frames": {
            "get": {
                "description": "Получение кадров видео по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Получение кадров видео",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id видео",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "processed",
                            "default"
                        ],
                        "type": "string",
                        "description": "Тип кадров (processed - обработанные, иначе - исходные)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список кадров видео",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Кадры видео не найдены",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Video": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "default = current timestamp",
                    "type": "string"
                },
                "groupIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "description": "serial",
                    "type": "integer"
                },
                "processedSource": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "description": "default = \"processing\"",
                    "type": "string"
                },
                "title": {
                    "description": "unique",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "default = current timestamp",
                    "type": "string"
                }
            }
        },
        "model.VideoCreate": {
            "type": "object",
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.VideoGroupUpdate": {
            "type": "object",
            "required": [
                "action",
                "videoId"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "add",
                        "remove"
                    ]
                },
                "groupId": {
                    "type": "integer",
                    "minimum": 0
                },
                "videoId": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        }
    }
}